Autologin to simpleSAML/idp
Adds an administrative option to the SAML plugin to instruct the plugin
to automatically redirect the user to the simpleSAML/idP without
displaying the login form.  This is useful for the case where it is
expected that all users would use the SAML login and the moodle SAML
login page becomes an unnecessary extra step.

To enable access when needed to the moodle login page a parameter is
added to the login page to show the moodle login: /login/?saml=false.
This provides a login for any local accounts.